Fixed repeated testrail ids
